refactor(fs): extract file-handle open-flag helpers (855 → 797 + 89)
Targeted helper extraction rather than a full class split. The
`FileHandle` class is one cohesive implementation of `IFileHandle`
with every method tied to the per-instance `GLib.IOChannel`, so an
`install*Methods` split would just shuffle lines without exposing
meaningful module boundaries.

The file-level helpers concerning the open-flag conversion move to
a sibling module `packages/node/fs/src/file-handle-open.ts`:

  - resolveIOMode(flags)  Node numeric/string → fopen IOMode
  - openIOChannel(path, mode, creat)  with the create-if-missing
                                       retry for `'r+'` + ENOENT
  - mapOpenError(err, path)  GLib.FileError → NodeJS.ErrnoException
  - O_WRONLY / O_RDWR / O_CREAT / O_TRUNC / O_APPEND constants
  - type IOMode

Plus one new export caught while writing the module:

  - shouldCreate(flags)

  Consolidates the "should we create-if-missing?" check from the
  constructor's `(options.flags & O_CREAT) !== 0` numeric-only check
  into a routine that also handles the string aliases (`'w'`,
  `'w+'`, `'wx'`, `'wx+'`, `'a'`, `'a+'`, `'ax'`, `'ax+'`, `'as'`,
  `'as+'`). The previous numeric-only check silently failed to
  auto-create when the user passed string flags like
  `await fsPromises.open(path, 'w')` against a non-existent path
  with IOChannel mode `'r+'`. Surfaced as ENOENT instead of the
  expected create-then-truncate. The new helper closes that hole.

file-handle.ts is now 797 LoC (down from 855); its imports section
is noticeably less cluttered.

640 GJS fs tests still green; tsc + build clean.

STATUS.md: new Completed entry.
